[step4] refactor(racingcar) package분리
