+++ src/main/java/racingcar/RacingGameLauncher.java
@@ -2,6 +2,11 @@ package racingcar;
 
 import java.util.List;
 import java.util.Scanner;
+import racingcar.domain.Car;
+import racingcar.controller.RacingGame;
+import racingcar.util.Parser;
+import racingcar.view.InputView;
+import racingcar.view.ResultView;
 
 public class RacingGameLauncher {
