+++ src/main/java/racingcar/domain/Referee.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 import java.util.Comparator;
 import java.util.List;
