+++ src/main/java/racingcar/util/Parser.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.util;
 
 public class Parser {
