+++ src/main/java/racingcar/view/ResultView.java
@@ -1,6 +1,7 @@
-package racingcar;
+package racingcar.view;
 
 import java.util.List;
+import racingcar.domain.Car;
 
 public class ResultView {
